+++ pkg/engine/assert/binding.go
@@ -4,30 +4,20 @@ import (
 	"context"
 
 	"github.com/jmespath-community/go-jmespath/pkg/binding"
-	"github.com/kyverno/kyverno-json/pkg/apis/v1alpha1"
 	"github.com/kyverno/kyverno-json/pkg/engine/template"
 	"k8s.io/apimachinery/pkg/util/validation/field"
 )
 
-func NewContextBindings(bindings binding.Bindings, value any, entries ...v1alpha1.ContextEntry) binding.Bindings {
-	var path *field.Path
-	path = path.Child("context")
-	for i, entry := range entries {
-		bindings = bindings.Register("$"+entry.Name, NewContextBinding(path.Index(i), bindings, value, entry))
-	}
-	return bindings
-}
-
-func NewContextBinding(path *field.Path, bindings binding.Bindings, value any, entry v1alpha1.ContextEntry) binding.Binding {
+func NewContextBinding(path *field.Path, bindings binding.Bindings, value any, entry any) binding.Binding {
 	return template.NewLazyBinding(
 		func() (any, error) {
-			expression := parseExpression(context.TODO(), entry.Variable.Value)
+			expression := parseExpression(context.TODO(), entry)
 			if expression != nil && expression.engine != "" {
 				if expression.foreach {
-					return nil, field.Invalid(path.Child("variable"), entry.Variable.Value, "foreach is not supported in context")
+					return nil, field.Invalid(path.Child("variable"), entry, "foreach is not supported in context")
 				}
 				if expression.binding != "" {
-					return nil, field.Invalid(path.Child("variable"), entry.Variable.Value, "binding is not supported in context")
+					return nil, field.Invalid(path.Child("variable"), entry, "binding is not supported in context")
 				}
 				projected, err := template.Execute(context.Background(), expression.statement, value, bindings)
 				if err != nil {
@@ -35,7 +25,7 @@ func NewContextBinding(path *field.Path, bindings binding.Bindings, value any, e
 				}
 				return projected, nil
 			}
-			return entry.Variable.Value, nil
+			return entry, nil
 		},
 	)
 }
